+++ pyxelrest/__init__.py
@@ -1,16 +1,8 @@
-import os
 from pyxelrest.version import __version__
 from pyxelrest._exceptions import PyxelRestException
 
 
 GENERATE_UDF_ON_IMPORT = True
-
-SERVICES_CONFIGURATION_FILE_PATH = os.path.join(
-    os.getenv("APPDATA"), "pyxelrest", "configuration", "services.yml"
-)
-LOGGING_CONFIGURATION_FILE_PATH = os.path.join(
-    os.getenv("APPDATA"), "pyxelrest", "configuration", "logging.yml"
-)
 
 
 def load(config: dict):