+++ pyxelrest/_generator_config.py
@@ -1,0 +1,8 @@
+import os
+
+SERVICES_CONFIGURATION_FILE_PATH = os.path.join(
+    os.getenv("APPDATA"), "pyxelrest", "configuration", "services.yml"
+)
+LOGGING_CONFIGURATION_FILE_PATH = os.path.join(
+    os.getenv("APPDATA"), "pyxelrest", "configuration", "logging.yml"
+)