+++ tests/test_no_services_configurations.py
@@ -1,11 +1,11 @@
 import os
 from importlib import reload, import_module
 
-import pyxelrest
+import pyxelrest._generator_config
 
 
 def test_without_service_configuration_file(tmpdir):
-    pyxelrest.SERVICES_CONFIGURATION_FILE_PATH = os.path.join(
+    pyxelrest._generator_config.SERVICES_CONFIGURATION_FILE_PATH = os.path.join(
         tmpdir, "non_existing_file.yml"
     )
     reload(import_module("pyxelrest._generator"))